refactor: refactored hir to remove 'implementation' folder
