+++ crates/hir/src/components.rs
