+++ crates/hir/src/declarations.rs
