--- crates/hir/src/implementation/mod.rs
@@ -1,5 +1,0 @@
-pub mod components;
-pub mod declarations;
-pub mod expression;
-pub mod names;
-pub mod statements;
